+++ include/pirate/internals/FakeServer.hpp
@@ -38,40 +38,40 @@ struct FakeServer
     struct Params
     {
 	std::string server_name = "Server";
-	ssize_t num_iterations = 20;
+	long num_iterations = 20;
 	bool use_hugepages = true;
 	
-	ssize_t nbytes_h2h = 0;  // memcpy host->host
-	ssize_t nbytes_h2g = 0;  // memcpy host->gpu (per GPU)
-	ssize_t nbytes_g2h = 0;  // memcpy host->gpu (per GPU)
-	ssize_t memcpy_blocksize = 0;
+	long nbytes_h2h = 0;  // memcpy host->host
+	long nbytes_h2g = 0;  // memcpy host->gpu (per GPU)
+	long nbytes_g2h = 0;  // memcpy host->gpu (per GPU)
+	long memcpy_blocksize = 0;
 	
-	ssize_t nbytes_gmem_kernel = 0;  // use memory bandwidth on GPU, using a GPU kernel (not cudaMemcpyDeviceToDevice())
-	ssize_t gmem_kernel_blocksize = 0;
+	long nbytes_gmem_kernel = 0;  // use memory bandwidth on GPU, using a GPU kernel (not cudaMemcpyDeviceToDevice())
+	long gmem_kernel_blocksize = 0;
 	
-	ssize_t nbytes_per_ssd = 0;
-	ssize_t nthreads_per_ssd = 0;
-	ssize_t nwrites_per_file = 0;
-	ssize_t nbytes_per_file = 16L * 1024L * 1024L;
+	long nbytes_per_ssd = 0;
+	long nthreads_per_ssd = 0;
+	long nwrites_per_file = 0;
+	long nbytes_per_file = 16L * 1024L * 1024L;
 	std::vector<std::string> ssd_list;
 	
 	// We use one downsampling thread for each value of 'src_bit_depth'.
-	ssize_t nbytes_downsample = 0;
+	long nbytes_downsample = 0;
 	
 	// Network parameters.
-	ssize_t nconn_per_ipaddr = 0;    // tcp connections per ip address
+	long nconn_per_ipaddr = 0;    // tcp connections per ip address
 	std::vector<std::string> ipaddr_list;      // e.g { "10.1.1.2", "10.1.2.2" }
 	
 	// If specified, include a SleepyWorker which sleeps for the specified number of seconds.
 	// This is useful either if there are no workers (e.g. test_03_receive_data()), or if you want to rate-limit workers.
-	ssize_t sleep_usec = 0;
+	long sleep_usec = 0;
 	
 	// Networking options. I'll probably add more later (e.g. zero-copy TCP)
-	ssize_t recv_bufsize = 512 * 1024;
+	long recv_bufsize = 512 * 1024;
 	bool use_epoll = true;
 	
 	// This parameter just determines how often the receiver threads synchronize counters.
-	ssize_t network_sync_cadence = 16 * 1024 * 1024;  // bytes per ip address
+	long network_sync_cadence = 16 * 1024 * 1024;  // bytes per ip address
 	
 	// If negative, all GPUs will be used.
 	int ngpu = -1;
