+++ include/pirate/internals/cpu_downsample.hpp
@@ -9,10 +9,10 @@ namespace pirate {
 #endif
 
 
-extern void cpu_downsample(int src_bit_depth, const uint8_t *src, uint8_t *dst, ssize_t src_nbytes, ssize_t dst_nbytes);
+extern void cpu_downsample(int src_bit_depth, const uint8_t *src, uint8_t *dst, long src_nbytes, long dst_nbytes);
 
 // 'src_nbytes' must be a multiple of this.
-extern ssize_t cpu_downsample_src_bytes_per_chunk(int src_bit_depth);
+extern long cpu_downsample_src_bytes_per_chunk(int src_bit_depth);
 
 
 } // namespace pirate
